chatbox-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai
langchain-huggingface
langchain-deepseek<1.0
chromadb
sentence-transformers==2.7.0
transformers>=4.41,<4.52
